Parser: Add simple in-browser JS parser comparator
This patch introduces a stupid benchmark to compare two
different implementations of a parser for the Gutenberg
grammer which end up as Javascript meant to run in the
browser.

The purpose is to aid development and optimization of
parsers and to support a competitive and lively third-party
parser ecosystem, or to be used by core devs trying to
refine what becomes the final default parser.

This is meant to be served as a static HTML page and
performs naive benchmarks and gathers statistics about
random runs of each parser over preselected documents.
It starts by comparing the outputs to see if they are
structurally the same, then it runs random parses and
measures the time spent parsing random documents.